Update Basic configs
