+++ cypress-tests/cypress.config.js
@@ -1,11 +1,4 @@
 const { defineConfig } = require("cypress");
-const { rmdir } = require("fs");
-const fs = require("fs");
-const XLSX = require("node-xlsx");
-
-const pg = require("pg");
-const path = require("path");
-const pdf = require("pdf-parse");
 
 module.exports = defineConfig({
   execTimeout: 1800000,
@@ -16,117 +9,44 @@ module.exports = defineConfig({
   viewportWidth: 1440,
   viewportHeight: 960,
   chromeWebSecurity: false,
-  trashAssetsBeforeRuns: true,
 
   e2e: {
     setupNodeEvents(on, config) {
-      on("task", {
-        readPdf(pathToPdf) {
-          return new Promise((resolve) => {
-            const pdfPath = path.resolve(pathToPdf);
-            let dataBuffer = fs.readFileSync(pdfPath);
-            pdf(dataBuffer).then(function ({ text }) {
-              resolve(text);
-            });
-          });
-        },
-      });
-
-      on("task", {
-        readXlsx(filePath) {
-          return new Promise((resolve, reject) => {
-            try {
-              let dataBuffer = fs.readFileSync(filePath);
-              const jsonData = XLSX.parse(dataBuffer);
-              // jsonData= jsonData[0].data
-              resolve(jsonData[0]["data"].toString());
-            } catch (e) {
-              reject(e);
-            }
-          });
-        },
-      });
-
-      on("task", {
-        deleteFile(filePath) {
-          return new Promise((resolve, reject) => {
-            const fullPath = path.resolve(filePath);
-            if (fs.existsSync(fullPath)) {
-              fs.unlink(fullPath, (err) => {
-                if (err) {
-                  console.error("Failed to delete file:", fullPath, err);
-                  return reject(err);
-                }
-                console.log("Deleted file:", fullPath);
-                resolve(true);
-              });
-            } else {
-              console.log("File not found, skipping delete:", fullPath);
-              resolve(false);
-            }
-          });
-        },
-      });
-
-      on("task", {
-        deleteFolder(folderName) {
-          return new Promise((resolve, reject) => {
-            if (fs.existsSync(folderName)) {
-              rmdir(folderName, { maxRetries: 10, recursive: true }, (err) => {
-                if (err) {
-                  console.error(err);
-                  return reject(err);
-                }
-                return resolve(null);
-              });
-            } else {
-              return resolve(null);
-            }
-          });
-        },
-      });
-
-      on("task", {
-        dbConnection({ dbconfig, sql }) {
-          const client = new pg.Pool(dbconfig);
-          return client.query(sql).then(result => {
-            return result;
-          });
-        },
-      });
-
-      on('before:browser:launch', (browser = {}, launchOptions) => {
-        if (browser.name === 'chrome' && browser.isHeadless === false) {
-          launchOptions.args.push(
-            '--disable-features=AutofillAccountStorage,PasswordManager',
-            '--disable-save-password-bubble',
-            '--disable-password-generation',
-            '--disable-password-manager-reauthentication'
-          );
-        }
-        return launchOptions;
-      });
+      require("./cypress/config/tasks")(on);
+      require("./cypress/config/browserConfig")(on);
       require("@cypress/code-coverage/task")(on, config);
-      // return config;
-
       require("./cypress/plugins/index.js")(on, config);
       return config;
     },
-    experimentalRunAllSpecs: true,
+
     baseUrl: "http://localhost:8082",
     specPattern: "cypress/e2e/happyPath/**/*.cy.js",
-    downloadsFolder: "cypress/downloads",
-    numTestsKeptInMemory: 0,
+
+    testIsolation: true,
+    retries: {
+      runMode: 2,
+      openMode: 0
+    },
     redirectionLimit: 3,
-    trashAssetsBeforeRuns: true,
+
+    numTestsKeptInMemory: 0,
     experimentalMemoryManagement: true,
-    coverage: false,
-    codeCoverageTasksRegistered: false,
+
+    experimentalRunAllSpecs: true,
+    experimentalOriginDependencies: true,
+
+    downloadsFolder: "cypress/downloads",
+    trashAssetsBeforeRuns: true,
     video: false,
     videoUploadOnPasses: false,
+    screenshotOnRunFailure: true,
+    screenshotsFolder: 'cypress/screenshots',
+
+    coverage: false,
+    codeCoverageTasksRegistered: false,
+
     // experimentalStudio: true,
     // experimentalPromptCommand: true,
-
     // projectId: "ca6324a0-4210-4f7e-846a-71ca2766ca4",
   },
 });
